+++ src/main/java/org/terasology/audio/OpenALManager.java
@@ -33,7 +33,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.game.CoreRegistry;
 import org.terasology.logic.players.LocalPlayer;
-import org.terasology.logic.manager.AudioManager;
 import org.terasology.rendering.cameras.Camera;
 import org.terasology.rendering.world.WorldRenderer;
 
@@ -150,7 +149,7 @@ protected Sound createStreamingAudio(String name, URL source) {
 }       */
 
     @Override
-    protected boolean checkDistance(Vector3d soundSource) {
+    protected boolean checkDistance(Vector3f soundSource) {
         Camera camera = CoreRegistry.get(WorldRenderer.class).getActiveCamera();
 
         if (camera == null) {
