+++ src/main/java/org/terasology/audio/SoundSource.java
@@ -16,6 +16,7 @@
 package org.terasology.audio;
 
 import javax.vecmath.Vector3d;
+import javax.vecmath.Vector3f;
 
 public interface SoundSource {
 
@@ -110,14 +111,14 @@ public interface SoundSource {
      * @param pos
      * @return
      */
-    public SoundSource setPosition(Vector3d pos);
+    public SoundSource setPosition(Vector3f pos);
 
     /**
      * Returns sound position in space
      *
      * @return
      */
-    public Vector3d getPosition();
+    public Vector3f getPosition();
 
     /**
      * Set sound source velocity
@@ -126,14 +127,14 @@ public interface SoundSource {
      * @param velocity
      * @return
      */
-    public SoundSource setVelocity(Vector3d velocity);
+    public SoundSource setVelocity(Vector3f velocity);
 
     /**
      * Returns sound source velocity
      *
      * @return
      */
-    public Vector3d getVelocity();
+    public Vector3f getVelocity();
 
     /**
      * Set sound source direction in cartesian coordinates
@@ -141,14 +142,14 @@ public interface SoundSource {
      * @param direction
      * @return
      */
-    public SoundSource setDirection(Vector3d direction);
+    public SoundSource setDirection(Vector3f direction);
 
     /**
      * Returns sound source direction in cartesian coordinates
      *
      * @return
      */
-    public Vector3d getDirection();
+    public Vector3f getDirection();
 
     /**
      * Returns sound source pitch
